Batallas funcionando
Se crea una clase Batalla para que se pueda simular la misma, crear batallones, cargarlos, ver su estado y realizar las rondas.
Además, se registra cada hechizo lanzado por personaje y se evita que el personaje pueda lanzar el mismo hechizo en la misma ronda

+++ MagosVsMortifagos/src/magos/Mago.java
@@ -3,8 +3,13 @@ package magos;
 import clases.Personaje;
 import clases.TipoPersonaje;
 
+//Se pleantea que los magos tienen el siguiente puntaje de nivel mágico:
+//	-Autores: 10
+//	-Profesores: 5
+//	-Estudiantes: 3
 public abstract class Mago extends Personaje {
-    public Mago(TipoPersonaje tipo, String nombrePersonaje, int nivelMagia, int puntosVida, boolean armado, boolean protegido) {
-        super(tipo, nombrePersonaje, nivelMagia, puntosVida, armado, protegido);
-    }
+	public Mago(TipoPersonaje tipo, String nombrePersonaje, int nivelMagia, float puntosVida, boolean armado,
+			boolean protegido) {
+		super(tipo, nombrePersonaje, nivelMagia, puntosVida, armado, protegido);
+	}
 }
